Experience bars HG and Life, quest experience

+++ Anointed/Celibate.swift
@@ -21,7 +21,7 @@ class Celibate : Skill {
     
     override func applyPassiveTraits() {
         
-        UNIVERSE.theGame.player.historyWithGodExperience += 1.0/60.0/60.0
+        UNIVERSE.theGame.player.historyWithGodExperience += Int(1.0/60.0/60.0)
         
     }
